backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart
polars
duckdb==0.10.2
groq
google-generativeai
fastapi
uvicorn
python-multipart
duckdb
polars
pandas
sqlalchemy
psycopg2-binary
pymysql
groq
google-generativeai
python-dotenv